Improve test coverage on [mds-daily] (#172)
Refactor + improve tests

I moved a bunch of stuff around, so we should probably
also write some Cypress tests for this to make sure
we don't break anything.
