+++ packages/mds-daily/tests/request-handlers.test.ts
@@ -1,0 +1,25 @@
+import db from '@mds-core/mds-db'
+import Sinon from 'sinon'
+import { DailyApiRequest, DailyApiResponse } from '../types'
+import { getRawTripData } from '../request-handlers'
+
+describe('Request handlers', () => {
+  describe('getRawTripData()', () => {
+    it('handles a db read error in getRawTripData()', async () => {
+      Sinon.replace(db, 'readEvents', Sinon.fake.rejects('fake-db-error'))
+      const res: DailyApiResponse = {} as DailyApiResponse
+      res.status = (code: number) => {
+        return {
+          send: () => code
+        } as any
+      }
+      await getRawTripData(
+        {
+          params: { trip_id: 'fake-trip-id' }
+        } as DailyApiRequest,
+        res as DailyApiResponse
+      )
+      Sinon.restore()
+    })
+  })
+})
