+++ packages/mds-daily/tests/test.ts
@@ -34,7 +34,9 @@ import cache from '@mds-core/mds-cache'
 import { makeDevices } from '@mds-core/mds-test-data'
 import { ApiServer } from '@mds-core/mds-api-server'
 import { TEST1_PROVIDER_ID } from '@mds-core/mds-providers'
-import { api } from '../api'
+import Sinon from 'sinon'
+import { agencyMiddleware, api } from '../api'
+import { dbHelperFail } from '../request-handlers'
 
 /* eslint-disable-next-line no-console */
 const log = console.log.bind(console)
@@ -167,6 +169,49 @@ describe('Tests API', () => {
       })
   })
 
+  describe('Agency-specific middleware', () => {
+    it('calls next()', async () => {
+      const nextSpy = Sinon.spy()
+      /* eslint-disable @typescript-eslint/no-explicit-any */
+      await agencyMiddleware(
+        {
+          // skip most of the middleware body with /health path
+          path: '/health'
+        } as any,
+        {} as any,
+        nextSpy
+      )
+      /* eslint-enable @typescript-eslint/no-explicit-any */
+      test.assert(nextSpy.calledOnce)
+    })
+
+    it('rejects an unauthorized user', async () => {
+      const statusSpy = Sinon.spy(statusCode => {
+        return {
+          send: () => {
+            return statusCode
+          }
+        }
+      })
+      /* eslint-disable @typescript-eslint/no-explicit-any */
+      await agencyMiddleware(
+        {
+          // follow main code path & do auth logic
+          path: '/foobar'
+        } as any,
+        {
+          locals: {} /* leave claims undefined */,
+          status: statusSpy
+        } as any,
+        () => {
+          return null
+        }
+      )
+      /* eslint-enable @typescript-eslint/no-explicit-any */
+      test.assert(statusSpy.calledOnceWithExactly(401))
+    })
+  })
+
   it('verifies 8 total events, and 4 are non-conformant', done => {
     request
       .get('/admin/last_day_stats_by_provider')
@@ -210,5 +255,10 @@ describe('Tests API', () => {
         // providerTestObject1.hasProperty('num_telemetry')
         done(err)
       })
+  })
+
+  it('handles a db helper failure', async () => {
+    await dbHelperFail(new Error('fake-error'))
+    await dbHelperFail('fake-error-string')
   })
 })
